backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pypdf==5.1.0
langchain==0.3.14
langchain-groq==0.2.2
langchain-community==0.3.14
chromadb==0.6.0
python-dotenv==1.0.1
requests==2.32.3
tiktoken==0.8.0